+++ src/pydiverse/pipedag/materialize/cache.py
@@ -26,7 +26,9 @@ class ImperativeMaterializationState:
         self.auto_suffix_counter = itertools.count()
 
     def add_table_lookup(self, obj, table: Table):
-        self.assumed_dependencies.add(table)
+        table_without_assumed_dependencies = table.copy_without_obj()
+        table_without_assumed_dependencies.assumed_dependencies = None
+        self.assumed_dependencies.add(table_without_assumed_dependencies)
         self.object_lookup[id(obj)] = table
         self.table_ids.add(id(table))