+++ src/pydiverse/pipedag/materialize/cache.py
@@ -23,12 +23,16 @@ class ImperativeMaterializationState:
         # tasks.
         self.object_lookup: dict[int, Table] = {}
         self.table_ids: set[int] = set()
+        self.tables: list[Table] = []  # this is just to make sure the IDs in self.table_ids are not reused
         self.auto_suffix_counter = itertools.count()
 
     def add_table_lookup(self, obj, table: Table):
-        self.assumed_dependencies.add(table)
+        table_without_assumed_dependencies = table.copy_without_obj()
+        table_without_assumed_dependencies.assumed_dependencies = None
+        self.assumed_dependencies.add(table_without_assumed_dependencies)
         self.object_lookup[id(obj)] = table
         self.table_ids.add(id(table))
+        self.tables.append(table)
 
 
 @dataclass(frozen=True)