+++ src/pydiverse/pipedag/materialize/materializing_task.py
@@ -699,6 +699,7 @@ class MaterializationWrapper:
                     list(sorted(state.assumed_dependencies)) if len(state.assumed_dependencies) > 0 else []
                 )
                 _ = my_store.materialize_task(task, task_cache_info, table, disable_task_finalization=True)
+                assert table.stage is not None
                 if not return_nothing:
 
                     def get_return_obj(return_as_type):
@@ -734,6 +735,8 @@ class MaterializationWrapper:
                     # substitute imperatively materialized object references with
                     # their respective table objects
                     x = object_lookup[id(x)]
+                    assert isinstance(x, Table)
+                    assert x.stage is not None
                 if isinstance(x, (Table, RawSql)):
                     # fill assumed_dependencies for Tables that were not yet
                     # materialized