+++ docs/source/changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.12.6 (2025-XX-XX)
 - Feat: Automatically check cache-validity of polars and pandas DataFrame tasks marked as lazy
+- Fix: Hash table names as part of lazy query_hash
 
 ## 0.12.5 (2025-11-26)
 - Workaround snowflake sqlalchemy dialect to enable ExternalTableReference to other database