+++ docs/source/changelog.md
@@ -4,6 +4,8 @@
 - Feat: Automatically check cache-validity of polars and pandas DataFrame tasks marked as lazy
 - Fix: lazy=True tasks in 100% cache valid stage can still rename outputs. This did not work for
   stage_commit_technique=READ_VIEWS.
+- Fix: Fix hang when using many imperative materializations in a single task
+  (the implementation was not designed for this and storage size grew unsustainably)
 
 ## 0.12.5 (2025-11-26)
 - Workaround snowflake sqlalchemy dialect to enable ExternalTableReference to other database