+++ docs/source/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.12.6 (2025-XX-XX)
+## 0.12.6 (2025-12-10)
 - Feat: Automatically check cache-validity of polars and pandas DataFrame tasks marked as lazy
 - Fix: lazy=True tasks in 100% cache valid stage can still rename outputs. This did not work for
   stage_commit_technique=READ_VIEWS.